chore(anvil): tidy up `anvil::server` module (#12663)
chore(anvil): reorganize `anvil::server` module
