+++ crates/anvil/src/lib.rs
@@ -4,6 +4,7 @@
 #![cfg_attr(docsrs, feature(doc_cfg))]
 
 use crate::{
+    error::{NodeError, NodeResult},
     eth::{
         EthApi,
         backend::{info::StorageInfo, mem},
@@ -14,7 +15,6 @@ use crate::{
     },
     filter::Filters,
     logging::{LoggingManager, NodeLogLayer},
-    server::error::{NodeError, NodeResult},
     service::NodeService,
     shutdown::Signal,
     tasks::TaskManager,
@@ -51,6 +51,7 @@ pub use config::{
 
 mod hardfork;
 pub use alloy_hardforks::EthereumHardfork;
+mod error;
 /// ethereum related implementations
 pub mod eth;
 /// Evm related abstractions
