+++ crates/anvil/src/server/beacon/error.rs
